Prefer u64 or u32 over usize.
Using `usize` in types that get serialized can cause compatibility
issues between different architectures.

+++ linera-chain/src/inbox.rs
@@ -66,7 +66,7 @@ pub struct InboxStateView<C> {
 )]
 pub struct Cursor {
     height: BlockHeight,
-    index: usize,
+    index: u64,
 }
 
 #[derive(Error, Debug)]
